Integrate product-mangement and cart-management branches (2026-05-21)
## Branch: origin/product-mangement

### New ProductManagement module
4 entities: Category, Product, ProductMedia, Review
4 repositories: CategoryRepository, ProductRepository, ProductMediaRepository, ReviewRepository
3 services: CategoryService, ProductService, ReviewService
3 controllers: CategoryController, ProductController, ReviewController

Key design decisions:
- Category is nullable-store-scoped (store_id nullable) — supports both global
  (ADMIN-managed) and per-store categories
- Product uses Integer productId (SERIAL) matching the DB; InventoryService
  receives productId.longValue() to bridge to Long-keyed Inventory
- ProductService bootstraps an Inventory record (quantity=0) on every product
  create — ON DELETE CASCADE handles cleanup on delete
- ProductService.toResponse() reads availableQuantity from InventoryService
  (Redis → DB) so quantity is never stale

Conflicts resolved:
- Fixed StoreManagement→StoreMangement package import typo in ProductService
- Deleted thin stub Category/Product entities and repos in StoreMangement/;
  updated StoreService catalog methods to import from ProductManagement
- Removed /api/ prefix from all @RequestMapping (context path is /api/v1)
- Wrapped all controller responses in ApiResponse<T>
- Public catalog now filters isActive=true products

### schema.sql: products & catalog section replaced
- categories: added description TEXT; kept store_id nullable FK
- products: renamed price→base_price; dropped inventory INT column; added
  is_active, rating default, created_at, updated_at; store FK ON DELETE CASCADE
- product_media: added media_type, alt_text, NOT NULL constraints, ON DELETE CASCADE
- reviews: changed user_id FK → customer_id FK on customers table; added title,
  rating CHECK(1..5), UNIQUE(product_id, customer_id), ON DELETE CASCADE

---

## Branch: origin/cart-management

### New CartManagement module
3 entities: ShoppingCart, CartItem, Wishlist
3 repositories: ShoppingCartRepository, CartItemRepository, WishlistRepository
4 services: CartService, CheckoutService, WishlistService, CartCleanupScheduler
2 controllers: CartController, WishlistController

Key design decisions:
- One cart per customer (UNIQUE customer_id); cart expires after 7 days
- CartCleanupScheduler (cron 2AM daily) releases reserved stock from expired
  carts back to Inventory — @EnableScheduling already present on Application
- CartService snapshots price at add time (priceAtAdd) to protect buyers from
  merchant price changes during checkout flow
- CheckoutService performs atomic stock reservation: if any item fails to
  reserve, all previously reserved items are rolled back
- app.shipping.flat-rate and app.tax.rate are fully configurable via
  application.properties with safe defaults (25.00 EGP / 0.00)
- Wishlist is keyed by user_id (not customer_id) matching schema design

Conflicts resolved:
- Removed /api/ prefix from /api/cart and /api/wishlist controllers
- Wrapped all controller responses in ApiResponse<T>
- Ignored branch's stale inventory/store_settings table definitions

### schema.sql: cart & checkout section replaced
- Replaced old thin `carts` table with `shopping_carts` (customer FK, expiry)
- Replaced old thin `cart_items` with real version (shopping_cart FK,
  price_at_add snapshot, added_at)
- Added `wishlists` table (UNIQUE user+product, ON DELETE CASCADE both FKs)

### application.properties
- Added app.shipping.flat-rate and app.tax.rate config keys

+++ src/main/java/com/example/flowmerceproject/StorefrontCustomization/controller/PublicStorefrontController.java
@@ -42,7 +42,7 @@ public class PublicStorefrontController {
     @GetMapping("/{storeId}/products/{productId}")
     public ResponseEntity<ApiResponse<CatalogDTOs.ProductResponse>> getProduct(
             @PathVariable Integer storeId,
-            @PathVariable Long productId) {
+            @PathVariable Integer productId) {
         return ResponseEntity.ok(ApiResponse.ok(
                 storeService.getPublicProduct(storeId, productId)));
     }
